Callback URL is now static
Previously, the configured callback_url was set to be a callable object
that would take the channel name as an argument. This is no longer
necessary as the channel's name will be included in the JSON payload of
the request. We can just use the same URL for all callbacks. If we find
a need for it to be callable we could always make it optionally
callable, but I didn't want to add that complexity without a real world
use case.